Old background and content scripts
Back when I first tried writing this, I explored a couple ideas using
the `skipCreditsEnabled` and `wwwplayer.config.skip.credits.enabled`
properties initialised on Netflix pages.

The first idea was to use a content script to change these values from
`true` to `false`. However, I found that this didn't do anything, very
likely because the global configs are read at load time, and my script
didn't overwrite them early enough.

I tried setting `"run_at": "document_start"` in the `content_scripts`
manifest, but that didn't do anything either.

The next idea was to use the `webRequest` API in a background script to
rewrite these properties before the page was rendered in the browser.
The problem with that idea was that you're not allowed to modify the
response body using this API. You can only modify headers.

That's where I stopped.

Today, I picked up the project again, this time using Mitmproxy to
modify the response body and flip these two properties. The properties
were correctly set to false, but it turns out that didn't change the
windowed-video-credits behaviour. Only then did I realise that these
properties instead probably control whether the "Skip Credits" _button_
is displayed or not, and have nothing to do with the small-frame video.
This is indicated by the `skipCreditsDisplayDuration` property, which is
set to 6000, likely meaning that the button should display for 6
seconds.

I don't know why it took me so long, but the real answer is so simple
and obvious: use CSS instead. The video player is assigned different
classes once the credits start rolling. All we need to do is:

* Ensure these classes don't get applied, or their styles don't get
  applied
* Ensure the "Back to Browse" button doesn't show
* Ensure the video controls still display when hovering over the video
  (possibly by virtually clicking on the video)