Try removing `postplay` class from player element
Use a `MutationObserver` to get an event when the Netflix player's class
attribute changes. Want to try removing the `.postplay` class, which
makes the player reduce to a small frame.

This doesn't currently work because when the player is reduced, a second
player element is added to the DOM, and my reference no longer
corresponds to the right player.

Looks like I'll need to try adding a mutation observer on child
elements, and watch for a new player element being added.