Use `classList.remove()`
Turns out `classList` isn't an array, it's a `DOMTokenList`, and it
doesn't have an `includes()` method. Instead, use `contains()`.

`DOMTokenList` also doesn't have a `filter()`, but we can remove the
`.postplay` class much easier with the `remove()` method.

Turns out I didn't need to worry about a `childList` mutation observer
as I previously thought. This version works. Now that the class is
removed, the video no longer reduces to the small frame.