Reactivate player controls and hide UI
Click the player frame to reactivate player controls. Player controls
disappear once the credits get minimised.

The problem with clicking the player frame, though, is that the controls
UI becomes visible. We want the controls DOM elements to come back, but
stay hidden.

Change the CSS class to "inactive" to hide the UI. There is a problem
with this currently: jiggling the mouse over the video after the class
is changed doesn't make the controls visible. Need to move the cursor
out of the frame and back in to reset the visibility state.

Add a couple CSS styles to hide:

* The "Back to Browse" button that appears for a second in the top left
  corner of the page
* The "RATED …" label in the top left corner of the page