Convert from a web extension to a user script
Since it looks like everything's going to happen in the content script,
might as well make this a user script instead.

Build with Browserify in order to get a compiled JS file compatible with
browsers.

For some reason I'm currently getting an error complaining that the
`stylesheet` variable is `null`. Need to look into that.