controls.hide(): Hide controls immediately
Don't use `wait_element` to hide the controls. I had done that because I
thought that the element wasn't there after clicking the "Watch Credits"
button, and I had to wait for it before trying to hide it.

Turns out the reason why the credits weren't hiding was that I just
shadowed the `controls` variable. This waiting isn't necessary, and in
fact it's undesirable, because it causes the player controls to appear
for a second before being hidden instead of not appearing at all.