styles: Hide the white border that appears around credits
A white border appears around the frame while the credits are in
minimised mode. This is visible on the top and left sides of the frame.

Since we automatically click on the video, the white border only appears
for a second or less, but it's still visually jarring, all the more
because it shifts the position of the video.

Remove the white border, ensuring it never appears.