Hide distractions when Seamless controls are added
Seamless controls can be added during the end credits of a series
episode. When those controls are added, the "Back to Browse" button
appears as well as the mouse cursor. A transparent promo background also
appears.

Hide the "Back to Browse" button and promo background in CSS.

Add a new `seamless` module to handle hiding the mouse cursor. Listen
for the addition of the `.SeamlessControls--container` element, which
means Seamless controls were activated. At this point, force the cursor
to be hidden. Show it again when the user moves the mouse.